CHANGE. [SYNTH] - 파형합성 오실레이터 파라미터 자료형 변경

+++ core/include/synthesis/oscillator.h
@@ -6,13 +6,13 @@
 
 class synthesis_oscillator {
 protected: /* parameter transform */
-	static double ratio(uint64_t note);
-	static double frequency(uint64_t note);
-	static double time(uint64_t sample);
-	static double phase(uint64_t note, uint64_t sample);
+	static double ratio(double note);
+	static double frequency(double note);
+	static double time(double sample);
+	static double phase(double note, double sample);
 
 public: /* oscillator */
-	static double sin(uint64_t note, uint64_t sample);
-	static double saw(uint64_t note, uint64_t sample, double skew = 0.000E+0);
-	static double square(uint64_t note, uint64_t sample, double duty = 0.500E+0);
+	static double sin(double note, double sample);
+	static double saw(double note, double sample, double skew = 0.000E+0);
+	static double square(double note, double sample, double duty = 0.500E+0);
 };
