+++ core/include/synthesis/unison.h
@@ -3,14 +3,14 @@
 
 class synthesis_unison : public synthesis_abstract {
 protected: /* parameter */
-	uint64_t depth_ = 3;
-	uint64_t detune_ = 5;
+	double depth_ = 3.000E+0;
+	double detune_ = 5.000E+0;
 
 public: /* constructor */
 	synthesis_unison() {}
 	synthesis_unison(
-	    uint64_t depth,
-	    uint64_t detune)
+	    double depth,
+	    double detune)
 	    : depth_(depth), detune_(detune) {}
 
 protected: /* abstract */
